### 2025-08-01 #8698 @editorialbot
[PRE REVIEW]: better_launch: A better replacement for the ROS2 launch system
**Submitting author:** @ndahn (<a href="http://orcid.org/0000-0003-2262-2989">Nikolas Dahn</a>)
**Repository:** https://github.com/dfki-ric/better_launch
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.1
