### 2025-11-16 d49cbf0
- Fixed introspection function not extracting func arg defaults
