### 2025-11-25 #8958 @lwilson
## Review checklist for [@lwilson](https://github.com/lwilson)

 * [x] **Repository:** Is the source code for this software available at the https://github.com/dfki-ric/better_launch?
 * [x] **Contribution and authorship:** Has the submitting author ([@ndahn](https://github.com/ndahn)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
