### 2025-12-02 7bb7602
Fixed get_launchfunc_signature_from_file not handling argument defaults
