### 2025-12-10 d6f0da4
- Fixed some issues with launch args - Fixed launch arg defaults being ignored in some circumstances - Fixed launch arg default strings not being parsed correctly - Introduced a SETTINGS singleton - Introduced BetterLaunch.is_included - introspection.find_calling_frame can now search different frames - Added option to limit maximum length of screen messages - Added log level filters and options - Click override options now have proper help text - Some progress on declarative launch files
