+++ better_launch/declarative/declarative.py
@@ -5,7 +5,7 @@ import logging
 
 from better_launch import BetterLaunch
 from better_launch.wrapper import _exec_launch_func
-from better_launch.utils.better_logging import Colormode
+from better_launch.utils.settings import Colormode, SETTINGS
 from better_launch.utils.click import DeclaredArg
 
 from .toml_parser import load as load_toml
@@ -137,20 +137,24 @@ def _get_toml_args(toml: dict) -> list[DeclaredArg]:
 def launch_toml(
     path: str,
     launch_args: dict[str, str] = None,
-    eval_mode: Literal["full", "literal", "none"] = None,
     *,
-    ui: bool = None,
-    join: bool = None,
-    screen_log_format: str = None,
-    file_log_format: str = None,
-    colormode: Colormode = None,
-    manage_foreign_nodes: bool = None,
-    keep_alive: bool = None,
+    # These should largely mirror the launch_this decorator
+    eval_mode: Literal["full", "literal", "none"] = "literal",
     allow_kwargs: bool = None,
+    ui: bool = False,
+    colormode: Colormode = Colormode.DEFAULT,
+    print_limit: int = 0,
+    screen_log_level: str | int = None,
+    screen_log_format: str = None,
+    file_log_level: str | int = None,
+    file_log_format: str = None,
+    manage_foreign_nodes: bool = None,
+    join: bool = None,
+    keep_alive: bool = None,
 ) -> None:
     """Execute a TOML better_launch launchfile.
 
-    In better_launch TOML launch files, most tables will be `call tables`. A call table is a dict that has a `func` key referring to one of the public :py:class:`BetterLaunch` member functions. All other attributes will be treated as keyword arguments to that function. Call tables are executed in the order they appear in the launch file, and the result of calling their associated function will be stored under the call table's name.
+    In better_launch TOML launchfiles, most tables will be `call tables`. A call table is a dict that has a `func` key referring to one of the public :py:class:`BetterLaunch` member functions. All other attributes will be treated as keyword arguments to that function. Call tables are executed in the order they appear in the launch file, and the result of calling their associated function will be stored under the call table's name.
 
     For example:
 
@@ -192,18 +196,11 @@ def launch_toml(
     - if     -> execute only if condition is true
     - unless -> execute only if condition is false
 
+    Similar to :py:meth:`launch_this`, a TOML launchfile may specify various settings to configure the launch process. In particular, all of the *keyword-only* arguments to this function can be specified with a `bl_` prefix as global args. For example, to set the `screen_log_level` from your launchfile you could add `bl_screen_log_level = "warning"` in the global scope.
+
     Lastly, there are a couple of special keys that may be declared in the TOML:
     - `bl_toml_format`: the better_launch TOML parser version your launch file was written for. Set this if the format has changed and you don't want to update your launch file. The current version is :py:data:`toml_format_version`.
-    - `bl_eval_mode`: if and how `$(eval ...)` substitutions should be supported.
-    - `bl_ui`: default value for starting the UI
-    - `bl_join`: whether to join the processes better_launch starts
-    - `bl_colormode`: default colormode
-    - `bl_screen_log_format`: default terminal output format
-    - `bl_file_log_format`: default file log format
-    - `bl_manage_foreign_nodes`: whether to show foreign nodes in the UI
-    - `bl_keep_alive`: whether to keep running after the last node exits
-    - `bl_allow_kwargs`: whether additional launch arguments are allowed
-
+    
     Parameters
     ----------
     path : str
@@ -214,12 +211,8 @@ def launch_toml(
         How to treat `eval` substitutions.
     ui : bool, optional
[…]
