### 2025-12-10 63df3fa
Declarative: - Substitutions now handle spaces correctly - Added some comments to the substitutions parser - Skipped call tables now store None as their result - New example for TOML turtle sim
