### 2025-12-11 b76a237
- Rewrote the settings implementation, much cleaner now - Removed some redundant output
