+++ better_launch/elements/ros2_launch_wrapper.py
@@ -28,7 +28,7 @@ def _launchservice_worker(
     launchservice_args: list[Any],
     launch_action_queue: Queue,
     log_queue: Queue,
-    settings: settings._Settings,
+    config: settings._Settings,
 ) -> None:
     """This function will run in a child process and will not have access to any objects already in memory UNLESS they are passed to it as arguments. See the comments for further details."""
     # Makes it easier to tell what's going on in the process table
@@ -63,7 +63,7 @@ def _launchservice_worker(
     import launch
 
     # Synchronize the settings from the host process
-    settings.SETTINGS = settings
+    settings._SETTINGS = config
 
     # LaunchService is a little stubborn about log formatting and always prepends the node's
     # name and then appends the output format, but this also allows us to capture the actual
@@ -85,7 +85,7 @@ def _launchservice_worker(
     std_handler = RecordForwarder(
         # TODO should always mimic the main process formatter configuration regarding colors etc.
         PrettyLogFormatter(
-            settings.SETTINGS.screen_log_format,
+            config.screen_log_format,
             roslog_pattern=r"\[(?P<name>.+)] *" + ROSLOG_PATTERN_BL,
         )
     )
@@ -267,7 +267,7 @@ class Ros2LaunchWrapper(AbstractNode):
                 self.launchservice_args,
                 self._launch_action_queue,
                 self._process_log_queue,
-                settings.SETTINGS,
+                settings.Settings(),
             ),
             name=self.name,
             daemon=True,
