+++ better_launch/utils/settings.py
@@ -1,10 +1,11 @@
+from typing import Any
 import os
 import logging
-import enum
-from dataclasses import dataclass
+from enum import IntEnum
+from dataclasses import dataclass, fields, replace, asdict
 
 
-class Colormode(enum.IntEnum):
+class Colormode(IntEnum):
     # Color messages based on their severity and highlight the sources in one color
     DEFAULT = 0
 
@@ -14,7 +15,7 @@ class Colormode(enum.IntEnum):
     # Color messages only based on the logging source
     SOURCE = 2
 
-    # Don't color messages
+    # Don"t color messages
     NONE = 3
 
     # Give a different color to each severity and logging source
@@ -42,127 +43,149 @@ default_screen_format = "[{levelcolor_start}{levelname}{levelcolor_end}] [{sourc
 default_file_format = "[{levelname}] [{asctime}] {message}"
 
 
-# TODO make this a singleton somehow
-@dataclass
+@dataclass(frozen=True)
 class _Settings:
-    __ui: bool = False
-    __colormode: Colormode = Colormode.DEFAULT
-    __print_limit: int = 0
-    __screen_log_level: int = logging.INFO
-    __screen_log_format: str = default_screen_format
-    __file_log_level: int = logging.INFO
-    __file_log_format: str = default_file_format
+    ui: bool = False
+    colormode: Colormode = Colormode.DEFAULT
+    print_limit: int = 0
+    screen_log_level: int = logging.INFO
+    screen_log_format: str = default_screen_format
+    file_log_level: int = logging.INFO
+    file_log_format: str = default_file_format
 
-    __initialized = False
-
-    def _initialize(
-        self,
-        ui: bool,
-        colormode: Colormode,
-        print_limit: int,
-        screen_log_level: int,
-        screen_log_format: str,
-        file_log_level: int,
-        file_log_format: str,
-    ) -> None:
-        """For all overrides the priority is `CLI > ENV > launchfile`.
+    def __init__(self, **kwargs):
         """
-        if self.__initialized:
-            raise RuntimeError("Settings cannot be initialized again")
+        Initialize settings with priority: env vars > kwargs > defaults.
 
-        # Whether to start the TUI
-        env_ui = os.environ.get("BL_UI_OVERRIDE")
-        if env_ui:
-            self.__ui = (env_ui.lower() == "true")
-        else:
-            self.__ui = ui
+        Parameters
+        ----------
+        kwargs
+            Function-level overrides
+        """
+        for field in fields(self):
+            # Get environment variable value
+            key = f"BL_{field.name.upper()}"
+            val = self._get_env_value(key, field.type)
 
-        # Logging colormode
-        env_colormode = os.environ.get("BL_COLORMODE")
-        if env_colormode:
+            # Resolve priority: env > kwargs > default
+            if val is not None:
+                res = val
+            elif field.name in kwargs:
+                res = kwargs[field.name]
+            else:
+                # Get default value from field
+                res = (
+                    field.default
+                    if field.default is not field.default_factory
+                    else field.default_factory()
+                )
+
+            # Cannot use regular setattr in a frozen dataclass
+            object.__setattr__(self, field.name, res)
+
+    def _get_env_value(self, env_key: str, field_type: type) -> Any:
+        """Get and convert environment variable based on field type.
+
+        Parameters
+        ----------
+        env_key : str
+            Env variable key to get the value from.
+        field_type : type
+            Type to convert the env variable to.
+
+        Returns
+        -------
+        Any
+            The value of the env variable converted to the target type, or None if the env variable is not set.
+
+        Raises
+        ------
+            ValueError if the env variable value could not be converted to the target type.
+        """
+        env_str = os.environ.get(env_key)
+        if env_str is None:
[…]
