### 2025-12-11 50cdf01
Fixed some declarative args
