### 2025-12-11 601f320
- Moved declarative.py to main namespace - Implemented toml includes (I think?)
