### 2025-12-15 6b7c532
Minor cleanup
