### 2025-12-17 e824554
Setting up mkdocs
