+++ docs/source/better_launch.tui.rst
@@ -1,29 +1,0 @@
-better\_launch.tui package
-==========================
-
-Submodules
-----------
-
-better\_launch.tui.better\_tui module
--------------------------------------
-
-.. automodule:: better_launch.tui.better_tui
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-better\_launch.tui.footer\_menu module
---------------------------------------
-
-.. automodule:: better_launch.tui.footer_menu
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: better_launch.tui
-   :members:
-   :undoc-members:
-   :show-inheritance:
