+++ docs/index.md
@@ -1,0 +1,21 @@
+# better_launch
+
+```
+                *       ,:
+    +                 ,' |
+           +         /   :
+       *          --'   /
++                \\/ /:/
+            *     / ://_\\
+       +       __/   /
+  -            )'-. /
+               ./  :\\
+        *       /.' '
+              '/'
+   '          +
+           .-"-
+          (    )
+       . .-'  '.
+      ( (.   )8:
+  .' _  / (_  ) '._
+```
