+++ CONTRIBUTING.md
@@ -55,4 +55,4 @@ in advance with deprecation warnings.
 ## Funding
 *better_launch* was initiated and is currently developed at the [Robotics Innovation Center](http://robotik.dfki-bremen.de/en/startpage.html) of the [German Research Center for Artificial Intelligence (DFKI)](http://www.dfki.de) in Bremen.
 
-![dfki-logo](media/dfki.png)
+![dfki-logo](docs/media/dfki.png)
