+++ mkdocs.yml
@@ -1,0 +1,71 @@
+site_name: better_launch
+repo_url: https://github.com/dfki-ric/better_launch
+copyright: Copyright &copy; 2025 DFKI GmbH
+theme:
+  favicon: media/logo.png
+  name: material
+  font:
+    text: Noto Sans
+  features:
+    - content.code.copy
+    - header.autohide
+    - navigation.tabs
+    - navigation.expand
+    - navigation.footer
+    - navigation.indexes
+    - navigation.instant
+    - navigation.tabs
+    - navigation.top
+    - search.highlight
+    - search.share
+    - search.suggest
+  palette:
+    - scheme: slate
+      primary: black
+      accent: indigo
+
+nav:
+  - Home: index.md
+  - Installation: installation.md
+  - Usage: usage.md
+  - Examples: examples.md
+  - Documentation: $api/better_launch.***
+
+plugins:
+  - group:
+      enabled: !ENV MKDOCS_MATERIAL_INSIDERS_ENABLED
+      plugins:
+          - optimize
+  - search
+  - mkapi
+
+markdown_extensions:
+  - abbr
+  - admonition
+  - attr_list
+  - def_list
+  - footnotes
+  - md_in_html
+  - toc:
+      toc_depth: "1-3"
+      permalink: true
+  # Python Markdown Extensions
+  - pymdownx.betterem:
+      smart_enable: all
+  - pymdownx.caret
+  - pymdownx.details
+  - pymdownx.highlight
+  - pymdownx.inlinehilite
+  - pymdownx.keys
+  - pymdownx.mark
+  - pymdownx.smartsymbols
+  - pymdownx.details
+  - pymdownx.superfences:
+      custom_fences:
+        - name: mermaid
+          class: mermaid
+          format: !!python/name:pymdownx.superfences.fence_code_format
+  - pymdownx.tilde
+  - pymdownx.emoji:
+      emoji_index: !!python/name:material.extensions.emoji.twemoji
+      emoji_generator: !!python/name:material.extensions.emoji.to_svg
