### 2025-12-22 603eb92
- bl will no longer reject launch files given by path if they don't have a known extension - group members and components in toml launchfiles can now be specified as subtables
