### 2025-12-22 5f3a870
Added toml and configuration pages
