+++ docs/howto/toml.md
@@ -1,1 +1,110 @@
 # :page_with_curl: TOML launchfiles
+
+The discussions over at ros discourse revealed that there was a need for launchfiles that are not code-based but can be parsed offline. TOML launchfiles are *better_launch*'s answer, based on the TOML format. They provide the full power of *better_launch* while limiting the complexity that would be achievable through python.
+
+This page will give you a general overview. Have a look at the [examples](../examples.md) for more details!
+
+## Call Tables
+In TOML launchfiles most tables are *call tables*. A call table is a dictionary with a `func` key referring to one of the
+public [BetterLaunch](../../reference/better_launch/launcher/#better_launch.launcher.BetterLaunch) member functions. Most other attributes are treated as keyword arguments to that function. Call tables execute in order of appearance, and their return value are stored under the table’s name.
+
+For example, the following call table will call [bl.find](../../reference/better_launch/launcher/#better_launch.launcher.BetterLaunch.find), pass `better_launch` as the package and `cube.sdf` as the filename, and store the returned value under `a_simple_cube`.
+
+```toml
+[a_simple_cube]
+func = "find"
+package = "better_launch"
+filename = "cube.sdf"
+```
+
+???+ warning
+
+    Every call table *must* have a unique name even if the call doesn't have a result or the result is not used!
+
+## Substitutions
+Substitutions take heavy inspiration from those found in ROS1 and should be familiar to many. However, as the TOML launchfile format is much more powerful, only the following substitutions were 
+deemed necessary for now:
+
+  - `${<K>}`: value of launch argument or call table `<K>`
+  - `${param <N> <P>}`: query node `<N>` for its ROS2 parameter `<P>`
+  - `${env <E> <D>}`: environment variable `<E>` (default `<D>`)
+  - `${eval <X>}`: evaluate Python expression `<X>`
+
+Substitutions may be nested (inner ones resolve first). The following call table will log the path of the cube we previously located:
+
+```toml
+[print_cube]
+func = "log"
+severity = "info"
+message = "A simple cube can be found at ${a_simple_cube}"
+```
+
+???+ note
+
+    By default `eval` will use `ast.literal_eval`. This can be changed by creating a global launch argument `bl_eval_mode` and setting it to either `full` (regular `eval`) or `none` (return expression without evaluation).
+
+## Conditions
+Call tables can be made conditional by adding an `if` or `unless` parameter. The table will only run if this value evaluates to true (false for `unless`) according to python truthiness. Be aware that non-empty strings are also considered true. If a call table doesn't run, `None` will be stored under its name.
+
+## Launch Arguments
+Launch arguments can be defined as global parameters before the first table. Their values can be used in substitutions just like any call table result. In addition, you may set the `bl_allow_kwargs` so arbitrary launch arguments can be passed to the launchfile (i.e. for nodes that consume them additional inputs).
+
+In order to support required launch arguments, *better_launch* makes a small extension to TOML: when defining your launch arguments, instead of assigning them a value you may also assign them one of python's primitive types *without quotes* (e.g. `my_arg = bool`). The launchfile will then require this argument to be passed and handle it as the specified type.
+
+## Function Context
+By default, call tables will lookup the function to run on the [BetterLaunch](../../reference/better_launch/launcher/#better_launch.launcher.BetterLaunch) instance. To run functions from the [convenience](convenience.md) or [gazebo](gazebo.md) modules, you may specify a `context` parameter and pass it either `convenience` or `gazebo`. This *cannot*  be used to run code from arbitrary modules - only these values are supported.
+
+For example, the following call table will start a rosbag recording:
+
+```toml
[…]
