### 2025-12-25 9ee5c3d
Written content done for now, still need to update the docstrings
