### 2025-12-26 cfa38e0
Fixed docstrings for mkdocs
