+++ better_launch/elements/abstract_node.py
@@ -37,9 +37,9 @@ class AbstractNode:
         remaps : dict[str, str], optional
             Topic remaps for this node.
         params : str | dict[str, Any], optional
-            Node parameters. If a string is passed it will be lazy loaded with :py:meth:`BetterLaunch.find`.
+            Node parameters. If a string is passed it will be lazy loaded with [BetterLaunch.find][].
         output : LogSink | Iterable[LogSink] | Iterable[str] | str, optional
-            Determines if and where this node's output should be directed. Common choices are `screen` to print to terminal, `log` to write to a common log file, `own_log` to write to a node-specific log file, and `none` to not write any output anywhere. See :py:meth:`configure_logger` for details.
+            Determines if and where this node's output should be directed. Common choices are `screen` to print to terminal, `log` to write to a common log file, `own_log` to write to a node-specific log file, and `none` to not write any output anywhere. See [configure_logger][] for details.
 
         Raises
         ------
@@ -107,7 +107,7 @@ class AbstractNode:
 
     @property
     def params(self) -> dict[str, Any]:
-        """The ROS params that were passed to this node. If a string was passed it is assumed to be a filepath and will be loaded with :py:meth:`BetterLaunch.find`."""
+        """The ROS params that were passed to this node. If a string was passed it is assumed to be a filepath and will be loaded with [BetterLaunch.find][]."""
         if isinstance(self._params, str):
             from better_launch import BetterLaunch
 
@@ -200,11 +200,11 @@ class AbstractNode:
         raise NotImplementedError()
 
     def start(self) -> None:
-        """Start this node. Once this succeeds, :py:meth:`is_running` will return True."""
+        """Start this node. Once this succeeds, [is_running][] will return True."""
         raise NotImplementedError()
 
     def shutdown(self, reason: str, signum: int = signal.SIGTERM, timeout: float = 0.0) -> None:
-        """Shutdown this node. Once this succeeds, :py:meth:`is_running` will return False.
+        """Shutdown this node. Once this succeeds, [is_running][] will return False.
 
         Parameters
         ----------
@@ -264,11 +264,11 @@ class AbstractNode:
             return None
 
     def is_lifecycle_node(self, timeout: float = 0.0) -> bool:
-        """Checks if this is a lifecycle node and initializes a : py:class:`LifecycleManager` if supported and not done so before.
+        """Checks if this is a lifecycle node and initializes a : py[LifecycleManager][] if supported and not done so before.
 
-        Note that if you simply want to check whether this node supports lifecycle management right now, check whether :py:meth:`lifecycle` is None will be considerably cheaper.
+        Note that if you simply want to check whether this node supports lifecycle management right now, check whether [lifecycle][] is None will be considerably cheaper.
 
-        Whether a node supports lifecycle management can only be known from outside once its process is started and it has registered with ROS. When this is called while the node is alive and it supports lifecycle management, a :py:class:`LifecycleManager` object will be initialized for it. This will persist even if the node is shutdown, but will obviously no longer provide useful functionality.
+        Whether a node supports lifecycle management can only be known from outside once its process is started and it has registered with ROS. When this is called while the node is alive and it supports lifecycle management, a [LifecycleManager][] object will be initialized for it. This will persist even if the node is shutdown, but will obviously no longer provide useful functionality.
 
[…]
