+++ better_launch/declarative.py
@@ -191,7 +191,7 @@ def launch_toml(
 ) -> None:
     """Execute a TOML better_launch launchfile.
 
-    In better_launch TOML launchfiles, most tables will be `call tables`. A call table is a dict that has a `func` key referring to one of the public :py:class:`BetterLaunch` member functions. All other attributes will be treated as keyword arguments to that function. Call tables are executed in the order they appear in the launch file, and the result of calling their associated function will be stored under the call table's name.
+    In better_launch TOML launchfiles, most tables will be `call tables`. A call table is a dict that has a `func` key referring to one of the public [BetterLaunch][] member functions. All other attributes will be treated as keyword arguments to that function. Call tables are executed in the order they appear in the launch file, and the result of calling their associated function will be stored under the call table's name.
 
     For example:
 
@@ -219,7 +219,7 @@ def launch_toml(
     eval_mode : Literal[&quot;full&quot;, &quot;literal&quot;, &quot;none&quot;], optional
         How to treat `eval` substitutions.
     ui : bool, optional
-        Whether to start the better_launch TUI. Superseded by the `BL_UI_OVERRIDE` environment variable and the `--bl_ui_override` argument.
+        Whether to start the better_launch TUI. Superseded by the `BL_UI` environment variable and the `--bl_ui_override` argument.
     allow_kwargs : bool, optional
         Whether additional launch arguments are allowed.
     colormode : Colormode, optional
@@ -229,17 +229,17 @@ def launch_toml(
         * source: one color per message source, don't colorize log severity
         * none: don't colorize anything
         * rainbow: colorize log severity and give each message source its own color
-        Superseded by the `BL_COLORMODE_OVERRIDE` environment variable and the `--bl_colormode_override` argument.
+        Superseded by the `BL_COLORMODE` environment variable and the `--bl_colormode_override` argument.
     print_limit : int, optional
         Limit the length of messages printed to the screen.
     screen_log_level : str | int, optional
         The minimum level for log messages to be printed to the terminal/screen. Can be either  "info", "warning", "error", "critical", or an arbitrary integer (e.g. logging.WARNING).
     screen_log_format : str, optional
-        Customize how log output will be formatted when printing it to the screen. Will be overridden by the `BL_SCREEN_LOG_FORMAT_OVERRIDE` environment variable. See :py:class:`PrettyLogFormatter` for details.
+        Customize how log output will be formatted when printing it to the screen. Will be overridden by the `BL_SCREEN_LOG_FORMAT` environment variable. See [PrettyLogFormatter][utils.better_logging.PrettyLogFormatter] for details.
     file_log_level : str | int, optional
         The minimum level for log messages to be written to the lot file. Can be either  "info", "warning", "error", "critical", or an arbitrary integer (e.g. logging.WARNING).
     file_log_format : str, optional
-        Customize how log output will be formatted when writing it to a file. Will be overridden by the `BL_FILE_LOG_FORMAT_OVERRIDE` environment variable. See :py:class:`PrettyLogFormatter` for details.
+        Customize how log output will be formatted when writing it to a file. Will be overridden by the `BL_FILE_LOG_FORMAT` environment variable. See [PrettyLogFormatter][utils.better_logging.PrettyLogFormatter] for details.
     manage_foreign_nodes : bool, optional
         If True, the TUI will also include node processes not started by this process. Has no effect if the TUI is not started.
     join : bool, optional
