+++ mkdocs.yml
@@ -48,7 +48,19 @@ nav:
 
 plugins:
   - search
-  - mkdocstrings
+  - mkdocstrings:
+      handlers:
+          python:
+            options:
+              docstring_style: numpy
+              show_signature_annotations: true
+              separate_signature: true
+              show_source: false
+              show_symbol_type_heading: true
+              show_symbol_type_toc: true
+              scoped_crossrefs: true
+              preload_modules:
+              - inspect
   - api-autonav:
       modules: ["better_launch"]
       nav_item_prefix: ""

[…]
