### 2025-12-26 e5915ce
Added DFKI logo to footer
