+++ docs/TODO.md
@@ -1,7 +1,4 @@
 # TODO
-- css for API reference
-- cleanup docstrings
-- add DFKI logo to footer
 - improve hero spaceship
 - social media links?
 - check social card
