### 2025-12-26 91472c1
Documented TUI shortcuts overrides
