+++ docs/howto/tui.md
@@ -12,6 +12,10 @@ See the single line of shortcuts at the bottom of each example? That's the TUI, 
 - changing the log level
 - etc.
 
+???+ tip
+
+    It is also possible to change the key bindings. The [settings page](settings.md#tui-shortcuts) contains further details.
+
 ```bash
 # Run this line to see it in action!
 bl better_launch 02_ui.launch.py
@@ -19,4 +23,6 @@ bl better_launch 02_ui.launch.py
 
 The TUI is also able to manage nodes started from different shells and processes, even if they have been started by ROS2 or other means. To do so, pass the `manage_foreign_nodes` flag to the wrapper or command line. Be aware though that this will not capture their output - to get their output you will have to use the *takeover* action from the TUI, which will restart the node process with the original arguments.
 
-> Foreign node processes are identified by having one of the following parameters in their arguments: `__ns`, `__name`, `__node`, `--ros-args`. This is always true for nodes started from launch files, but fails when they were started via `ros2 run` or other means. As far as I'm aware, there is no better way right now.
+???+ note
+
+    Foreign node processes are identified by having one of the following parameters in their arguments: `__ns`, `__name`, `__node`, `--ros-args`. This is always true for nodes started from launch files, but fails when they were started via `ros2 run` or other means. As far as I'm aware, there is no better way right now.
