### 2026-01-12 2ca1c42
Added missing write permission to github workflow
