### 2026-01-12 0c6548a
More explicit handling of node names
