### 2026-01-12 d872068
feat: add explicit node parameter overrides via CLI
