+++ better_launch/launcher.py
@@ -117,6 +117,7 @@ class BetterLaunch(metaclass=_BetterLaunchMeta):
 
     _launchfile: str = None
     _launch_func_args: dict[str, Any] = {}
+    _node_param_overrides: list[tuple[str, str, str]] = []
 
     def __init__(
         self,
@@ -243,7 +244,9 @@ Takeoff in 3... 2... 1...
             except (CancelledError, TimeoutError):
                 pass
 
-        self.logger.critical(f"Reminder: log files are at {roslog.launch_config.log_dir}")
+        self.logger.critical(
+            f"Reminder: log files are at {roslog.launch_config.log_dir}"
+        )
 
     def get_unique_name(self, name: str = "", check_running_nodes: bool = True) -> str:
         """Returns a unique name. If a name is provided it will be prepended with an underscore.
@@ -336,10 +339,10 @@ Takeoff in 3... 2... 1...
             nodes.append(self._ros2_launcher)
 
         if include_foreign:
-            # Note that self.shared_node.get_node_names_and_namespaces() will only give us the 
+            # Note that self.shared_node.get_node_names_and_namespaces() will only give us the
             # names and namespaces, but no handle on the actual processes, not even a package
 
-            # TODO should check if it's a composer and has subnodes, but we won't know the 
+            # TODO should check if it's a composer and has subnodes, but we won't know the
             # components' handles or plugins...
             # if include_components:
             #     for f in foreign:
@@ -347,7 +350,7 @@ Takeoff in 3... 2... 1...
             #             composer = Composer(f)
             #             for cid, component in composer.get_live_components().items():
             #                 nodes.append(Component(...))
-            
+
             foreign = self.get_foreign_nodes()
             nodes.extend(foreign)
 
@@ -667,9 +670,7 @@ Takeoff in 3... 2... 1...
             If `package` contains path separators, or if a `filename` is provided but could not be found within base path.
         """
         if filename and os.path.isabs(filename):
-            self.logger.info(
-                f"find({package}, {filename}, {subdir}):1 -> {filename}"
-            )
+            self.logger.info(f"find({package}, {filename}, {subdir}):1 -> {filename}")
             return filename
 
         if not package:
@@ -717,7 +718,6 @@ Takeoff in 3... 2... 1...
         raise ValueError(
             f"Could not find file or directory (package={package}, filename={filename}, subdir={subdir}), searched path was {base_path}"
         )
-
 
     def load_params(
         self,
@@ -1465,6 +1465,31 @@ Takeoff in 3... 2... 1...
         if hidden and not name.startswith("_"):
             name = "_" + name
 
+        # Check for parameter overrides
+        if self._node_param_overrides:
+            overrides = [(p, v) for n, p, v in self._node_param_overrides if n == name]
+
+            if overrides:
+                # Prepare params for modification
+                if params is None:
+                    params = {}
+                elif isinstance(params, str):
+                    params = self.load_params(configfile=params)
+                else:
+                    # Assumed dict
+                    params = params.copy()
+
+                for param_name, value in overrides:
+                    try:
+                        parsed_value = yaml.safe_load(value)
+                    except Exception:
+                        parsed_value = value
+
+                    self.logger.info(
+                        f"Overriding param '{param_name}' for node '{name}' with '{parsed_value}'"
+                    )
+                    params[param_name] = parsed_value
+
         group = self.group_tip
         namespace = group.assemble_namespace()
 
@@ -1637,7 +1662,14 @@ Takeoff in 3... 2... 1...
             else:
                 raise ValueError(f"Unknown container mode '{variant}")
 
-            node_ref = Node(package, executable, name, namespace, remaps=component_remaps, output=output)
+            node_ref = Node(
+                package,
[…]
