### 2026-01-16 #8958 @diehlpk
@ndahn could you please work on the community guidelines? I would like to get this paper published by end of the month. We have quite a backlog and I would like to edit another paper.
