+++ docs/paper.bib
@@ -1,3 +1,11 @@
+@misc{ros2_launch_design,
+  author = {William Woodall},
+  title = {ROS 2 Launch System},
+  year = {2019},
+  publisher = {Open Source Robotics Foundation},
+  journal = {ROS 2 Design},
+  url = {https://design.ros2.org/articles/roslaunch.html}
+}
 @misc{launchmap,
   author = {Sakshay Mahna and Luana Priscila da Costa Silva},
   title = {LaunchMap – Visualize Your ROS 2 Launch Files},
@@ -37,4 +45,36 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/xqms/rosmon}
+}
+@misc{ros2_substitution_example,
+  author = {Open Robotics},
+  title = {Using Substitutions},
+  year = {2025},
+  publisher = {Open Robotics},
+  journal = {ROS 2 Tutorials},
+  url = {https://docs.ros.org/en/jazzy/Tutorials/Intermediate/Launch/Using-Substitutions.html}
+}
+@misc{py-spy,
+  author = {Ben Frederickson},
+  title = {py-spy},
+  year = {2025},
+  publisher = {Github},
+  journal = {Github repository},
+  url = {https://github.com/benfred/py-spy}
+}
+@misc{memray,
+  author = {Bloomberg Engineering},
+  title = {memray},
+  year = {2025},
+  publisher = {Github},
+  journal = {Github repository},
+  url = {https://github.com/bloomberg/memray}
+}
+@misc{psutil,
+  author = {Giampaolo Rodola},
+  title = {psutil},
+  year = {2025},
+  publisher = {Github},
+  journal = {Github repository},
+  url = {https://github.com/giampaolo/psutil}
 }
