### 2026-01-19 eb3aa3e
Added AI clause
