### 2026-01-20 #8958 @diehlpk
@ndahn Some minor comments:

1. Frederickson, 2025; Rodola, 2025).. -> Please remove on dot at the end of the sentence
2. ~~Can you please provide a link to the source code in the example?~~
