### 2026-01-25 c5e31f3
feat: enable explicit node parameter overrides
