+++ better_launch/wrapper.py
@@ -1,7 +1,9 @@
 from typing import Callable
-import os
-import platform
 from ast import literal_eval
+import os
+import sys
+import platform
+
 import signal
 import inspect
 import click
@@ -24,6 +26,8 @@ from better_launch.utils.click import (
     get_click_options,
     get_click_overrides,
     get_click_launch_command,
+    parse_node_params,
+    args_to_dict,
 )
 from better_launch.ros import logging as roslog
 
@@ -75,7 +79,7 @@ def launch_this(
     def decoration_helper(func):
         sig = inspect.signature(func)
         declared_args = _get_declared_args(sig, func.__doc__)
-        
+
         func_doc = doc.parse(func.__doc__)
 
         argspec = inspect.getfullargspec(func)
@@ -127,7 +131,9 @@ def _init_signal_handlers() -> None:
         signal.signal(signal.SIGQUIT, sigterm_handler)
 
 
-def _get_declared_args(signature: inspect.Signature, docstring: str = None) -> list[DeclaredArg]:
+def _get_declared_args(
+    signature: inspect.Signature, docstring: str = None
+) -> list[DeclaredArg]:
     # Extract more fine-grained information from the docstring
     param_docstrings = {}
     if docstring:
@@ -203,7 +209,11 @@ def _exec_launch_func(
         include_args: dict = glob[_bl_include_args]
         bl.logger.info(f"Including launch file: {includefile} (args={include_args})")
 
-        call_kw = {a.name: a.default for a in declared_args if a.default != DeclaredArg._undefined}
+        call_kw = {
+            a.name: a.default
+            for a in declared_args
+            if a.default != DeclaredArg._undefined
+        }
 
         for key, val in include_args.items():
             if allow_kwargs or key in call_kw:
@@ -275,31 +285,27 @@ def _exec_launch_func(
             overrides.colormode,
         )
 
-        if allow_kwargs is not None:
-            # If the launch func defines a **kwarg we can pass all extra arguments to it, with
-            # the caveat that these extra args need to be defined as `-[-]<key> val` tuples.
-            assert (
-                len(ctx.args) % 2 == 0
-            ), f"extra arguments need to be '--<key> <value>' tuples ({ctx.args})"
+        # Parse node-specific param overrides from ctx.args (extra unknown args)
+        # Convert ctx.args list to dict, then extract node params
+        extra_args_dict = args_to_dict(ctx.args)
+        remaining_kwargs, node_overrides = parse_node_params(extra_args_dict)
+        applied_overrides = node_overrides if overrides.node_param_override else {}
+        if applied_overrides:
+            BetterLaunch().set_node_param_overrides(applied_overrides)
 
-            for i in range(0, len(ctx.args), 2):
-                key = ctx.args[i]
-                if not key.startswith("-"):
-                    raise ValueError("Extra argument keys must start with a dash")
-
-                val = ctx.args[i + 1]
-                try:
-                    val = literal_eval(val)
-                except Exception:
-                    # Keep val as a string
-                    pass
-
-                kwargs[key.strip("-")] = val
+        if allow_kwargs and remaining_kwargs:
+            for key, value in remaining_kwargs.items():
+                if isinstance(value, str):
+                    try:
+                        value = literal_eval(value)
+                    except Exception:
+                        pass
+                kwargs[key] = value
 
         # By default BetterLaunch has access to all arguments from its launch function
         BetterLaunch._launch_func_args = dict(kwargs)
 
-        # Wrap the launch function so we can do some preparation and cleanup tasks. 
+        # Wrap the launch function so we can do some preparation and cleanup tasks.
         def launch_func_wrapper():
             try:
[…]

[…]
