### 2026-01-25 20c731c
Removed qualifier path from returned params
