### 2026-01-25 e8c5660
Fixed problem in load_params with nested dicts and lists
