### 2026-01-25 9b56b59
Removed keyword arg from exec to be compatible across more python versions
