+++ CONTRIBUTING.md
@@ -1,28 +1,20 @@
 # Contributing to better_launch
-Please inform the maintainer as early as possible about your planned
-feature developments, extensions, or bugfixes that you are working on.
-An easy way is to open an issue or a pull request in which you explain
-what you are trying to do.
+Please inform me as early as possible about your planned developments. You may have noticed that *better_launch* is somewhat opinionated and I will be picky about what to include :) An easy way is to open an issue in which you explain what you are trying to do.
 
 ## Pull Requests
-The preferred way to contribute to *better_launch* is to fork the
-[main repository](https://github.com/dfki-ric/better_launch) on GitHub, then submit a "pull request"
-(PR):
+The preferred way to contribute to *better_launch* is to fork the [repository](https://github.com/dfki-ric/better_launch) on GitHub *and work from the **devel** branch*, then submit a "pull request" (PR):
 
-1. [Create an account](https://github.com/signup/free) on
-   GitHub if you do not already have one.
+1. [Create an account](https://github.com/signup/free) on GitHub if you do not already have one.
 
-2. Fork the [project repository](https://github.com/dfki-ric/better_launch):
-   click on the 'Fork' button near the top of the page. This creates a copy of
-   the code under your account on the GitHub server.
+2. Fork the [project repository](https://github.com/dfki-ric/better_launch): click on the 'Fork' button near the top of the page. This creates a copy of the code under your account on the GitHub server.
 
 3. Clone this copy to your local disk:
 
-        $ git clone git@github.com:dfki-rc/better_launch.git
+        $ git clone -b devel git@github.com:dfki-rc/better_launch.git
 
 4. Create a branch to hold your changes:
 
-        $ git checkout -b my-feature
+        $ git checkout -b my-feature devel
 
     and start making changes. Never work in the ``master`` branch!
 
@@ -34,22 +26,17 @@ The preferred way to contribute to *better_launch* is to fork the
 
     to record your changes in Git, then push them to GitHub with::
 
-       $ git push -u origin my-feature
+        $ git push -u origin my-feature
 
-Finally, go to the web page of the your fork of the bolero repo,
-and click 'Pull request' to send your changes to the maintainers for review.
-request.
+Finally, go to the web page of the your fork of the bolero repo, and click 'Pull request' to send your changes to the maintainers for review request.
 
 ## Merge Policy
-Developers have to submit pull requests. Those will be reviewed by at least
-one other developer and merged by the maintainer. New features must be
-documented and tested. Breaking changes must be discussed and announced
-in advance with deprecation warnings.
+Developers have to submit pull requests. Those will be reviewed by at least one other developer and merged by the maintainer. New features must be documented and tested. Breaking changes must be discussed and announced in advance with deprecation warnings.
 
 ## Roadmap
 - a tool for generating launch graphs from *better_launch* launch files
 - more interactions for the TUI like setting live parameters
-- a tool for converting regular ROS2 launch files to *better_launch*. This is a big maybe as it seems quite annoying to do
+- something for converting regular ROS2 launch files to *better_launch*. Some people had good experiences using LLMs
 - who knows :)
 
 ## Funding
