### 2026-01-29 ac33c9e
Rewrote the readme for the new website
