### 2026-01-29 fe463eb
Removed unused arg
