### 2026-01-29 #8958 @ndahn
@diehlpk
I have released v`1.3.0` here: https://github.com/dfki-ric/better_launch/releases/tag/1.3.0
The Zenodo submission is under review by the community (my institute) and the DOI is `10.5281/zenodo.18414064`.
