### 2026-02-04 #8958 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
