### 2026-02-09 d7b10b8
minor changes in paper.md
