### 2026-02-09 #8958 @danielskatz
@ndahn - I've suggested changes for your paper in https://github.com/dfki-ric/better_launch/pull/54 - please merge this, or let me know what you disagree with, then we can proceed.

I also notice that you don't have an acknowledgments section, which is fine, but this is also a time you could add one if you wanted to.
