### 2026-02-09 #8958 @danielskatz
👋 @lwilson - I note that you did't check one of your review criteria, regarding community guidelines. Are you ok with this as in the paper/repo?  If so, can you check it?  If not, what do you think @ndahn still needs to do?
