### 2026-02-14 8dfe369
feat(launcher): add process() method for non-ROS processes
- Add BetterLaunch.process() method as clean API for ExecuteProcess-like functionality
- Supports cmd as str or list[str], auto-parsing via shlex.split()
- Resolves bare commands to absolute paths via shutil.which()
- Enforces raw=True and package="" internally
- Excludes ROS-specific kwargs (remaps, params, log_level, etc.)
- Add unit tests for method signature and docstring
