+++ better_launch/launcher.py
@@ -7,6 +7,8 @@ import inspect
 import time
 import threading
 import subprocess
+import shlex
+import shutil
 from fnmatch import fnmatch
 from pathlib import Path
 from concurrent.futures import Future, CancelledError, TimeoutError
@@ -805,10 +807,7 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
                     val = val.get("ros__parameters", val)
 
                     # Global parameters should always be included
-                    if (
-                        not path
-                        or fnmatch(path, qualifier)
-                    ):
+                    if not path or fnmatch(path, qualifier):
                         for param_name, param_val in val.items():
                             final_params[param_name] = param_val
 
@@ -1464,7 +1463,9 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
         if name is None:
             name = f"{package}_{executable}"
             if not anonymous:
-                self.logger.warning(f"Name of node {package}/{executable} not set, will use anonymous name")
+                self.logger.warning(
+                    f"Name of node {package}/{executable} not set, will use anonymous name"
+                )
                 anonymous = True
 
         if anonymous:
@@ -1511,6 +1512,103 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
                     node.lifecycle.transition(lifecycle_target)
 
         return node
+
+    def process(
+        self,
+        cmd: str | list[str],
+        name: str = None,
+        *,
+        env: dict[str, str] = None,
+        isolate_env: bool = False,
+        output: LogSink | Iterable[LogSink] | Iterable[str] | str = LogSink.SCREEN,
+        anonymous: bool = False,
+        hidden: bool = False,
+        on_exit: Callable = None,
+        max_respawns: int = 0,
+        respawn_delay: float = 0.0,
+        use_shell: bool = False,
+        autostart_process: bool = True,
+    ) -> Node:
+        """Starts an arbitrary process.
+
+        This method provides a clean API for starting non-ROS processes, similar to ROS2's
+        `ExecuteProcess`. It resolves bare command names to absolute paths using `shutil.which`.
+
+        Parameters
+        ----------
+        cmd : str | list[str]
+            The command to execute. If a string is provided, it will be split using `shlex.split`.
+        name : str, optional
+            The name of the process. If not provided, it will be derived from the command.
+        env : dict[str, str], optional
+            Additional environment variables to set for the process. The process will merge these with the environment variables of the better_launch host process unless `isolate_env` is True.
+        isolate_env : bool, optional
+            If True, the process' env will not be inherited from the parent process and only those passed via `env` will be used.
+        output : LogSink | Iterable[LogSink] | Iterable[str] | str, optional
+            Determines if and where this process' output should be directed. Defaults to `LogSink.SCREEN`.
+        anonymous : bool, optional
+            If True, the process name will be appended with a unique suffix to avoid name conflicts.
+        hidden : bool, optional
+            If True, the process name will be prepended with a "_", hiding it from common listings.
+        on_exit : Callable, optional
+            A function to call when the process terminates (after any possible respawns).
+        max_respawns : int, optional
+            How often to restart the process if it terminates.
+        respawn_delay : float, optional
+            How long to wait before restarting the process after it terminates.
+        use_shell : bool, optional
+            If True, invoke the executable via the system shell.
+        autostart_process : bool, optional
+            If True, start the process before returning from this function.
+
+        Returns
+        -------
+        Node
+            The node object wrapping the process.
+
+        Raises
+        ------
+        ValueError
+            If the command is empty.
+        FileNotFoundError
+            If the executable cannot be found.
+        """
+        if not cmd:
[…]
