### 2026-02-24 bef5055
- Much simpler and hopefully better implementation for load_params - AbstractNode._flat_params will now resolve wildcards
