### 2026-02-24 ca10ca4
Now using shlex to split cmds in exec
