### 2026-02-24 #8958 @danielskatz
@ndahn - please check the proof that this generates
