### 2026-02-25 fb67103
Figuring out different install file layout in jazzy
