+++ better_launch/launcher.py
@@ -691,9 +691,10 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
                 raise ValueError(
                     f"Package must be a single name, not a path ({package})"
                 )
+
             base_path = get_package_prefix(package)
         else:
-            base_path = os.getcwd()
+            raise ValueError(f"find({package}, {filename}, {subdir}): could not determine package")
 
         if not filename and subdir in (None, "", "**"):
             self.logger.info(f"find({package}, {filename}, {subdir}):2 -> {base_path}")
@@ -702,28 +703,50 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
         if not subdir:
             subdir = "**"
 
-        for candidate in Path(base_path).glob(subdir):
-            if not filename:
-                # Return the first candidate
-                ret = str(candidate.resolve().absolute())
-                self.logger.info(f"find({package}, {filename}, {subdir}):3 -> {ret}")
-                return ret
-
-            if candidate.is_file() and candidate.match(f"**/{filename}"):
-                # We found a match
-                ret = str(candidate.resolve().absolute())
-                self.logger.info(f"find({package}, {filename}, {subdir}):4 -> {ret}")
-                return ret
-
-            elif candidate.is_dir():
-                # Candidate is a dir, search the filename within
-                ret = next(candidate.glob(f"**/{filename}"), None)
-                if ret:
-                    ret = str(ret.resolve().absolute())
-                    self.logger.info(
-                        f"find({package}, {filename}, {subdir}):5 -> {ret}"
-                    )
+        def search_base_path(base_path: Path | str) -> str:
+            for candidate in Path(base_path).glob(subdir):
+                if not filename:
+                    # Return the first candidate
+                    ret = str(candidate.resolve().absolute())
+                    self.logger.info(f"find({package}, {filename}, {subdir}):3 -> {ret}")
                     return ret
+
+                if candidate.is_file() and candidate.match(f"**/{filename}"):
+                    # We found a match
+                    ret = str(candidate.resolve().absolute())
+                    self.logger.info(f"find({package}, {filename}, {subdir}):4 -> {ret}")
+                    return ret
+
+                elif candidate.is_dir():
+                    # Candidate is a dir, search the filename within
+                    ret = next(candidate.glob(f"**/{filename}"), None)
+                    if ret:
+                        ret = str(ret.resolve().absolute())
+                        self.logger.info(
+                            f"find({package}, {filename}, {subdir}):5 -> {ret}"
+                        )
+                        return ret
+            
+            return None
+
+        if self.ros_distro()[0].lower() >= "j":
+            # In jazzy and later package files are no longer collected in a singular package
+            # folder. Instead, workspace/install has global include, bin, lib, share, etc. 
+            # folders
+            ret = search_base_path(base_path + "/bin/")
+            print(f"### SEARCHING {base_path}/bin/")
+            if ret:
+                return ret
+
+            for install_dir in Path(base_path).glob(f"*/{package}/"):
+                print(f"### SEARCHING {install_dir}")
+                ret = search_base_path(install_dir)
+                if ret:
+                    return ret
+        else:
+            ret = search_base_path(base_path)
+            if ret:
+                return ret
 
         raise ValueError(
             f"Could not find file or directory (package={package}, filename={filename}, subdir={subdir}), searched path was {base_path}"
