+++ better_launch/elements/node.py
@@ -11,7 +11,6 @@ import subprocess
 import queue
 from pprint import pformat
 import json
-from ament_index_python.packages import get_package_prefix
 
 from better_launch.utils.better_logging import LogSink, ROSLOG_PATTERN_BL
 from .abstract_node import AbstractNode
@@ -155,12 +154,8 @@ class Node(AbstractNode, LiveParamsMixin):
             return
 
         try:
-            # In jazzy and later, package files are no longer collected in a singular package
-            # folder. Instead, workspace/install has global include, bin, lib, share, etc. 
-            # folders where all the files are placed, which is quite annoying for us
-            install_path = get_package_prefix(self.package)
-            cmd = launcher.find(install_path, self.executable, f"lib/**/{self.package}")
-            
+            cmd = launcher.find(self.package, self.executable, "lib/**")
+
             final_cmd = [cmd]
             if self.cmd_args:
                 final_cmd.extend(self.cmd_args)
@@ -185,14 +180,14 @@ class Node(AbstractNode, LiveParamsMixin):
                     qualifier = self.remap_qualifier
                     if not qualifier.endswith(":"):
                         qualifier += ":"
-                
+
                 for src, dst in self._ros_args().items():
                     if qualifier:
                         if src in ("__ns", "__node", "__name"):
                             src = qualifier + src
                         elif self.qualify_all_remaps and ":" not in src:
                             src = qualifier + src
-                    
+
                     final_cmd.extend(["-r", f"{src}:={dst}"])
 
             # If an env is specified ROS2 lets it completely replace the host env. We cover this
@@ -207,9 +202,7 @@ class Node(AbstractNode, LiveParamsMixin):
             final_cmd = [str(s) for s in final_cmd]
 
             env_str = pformat(self.env, compact=True)
-            self.logger.info(
-                f"Starting process '{' '.join(final_cmd)}', env={env_str}"
-            )
+            self.logger.info(f"Starting process '{' '.join(final_cmd)}', env={env_str}")
 
             # Start the node process
             self._process = subprocess.Popen(
@@ -288,8 +281,8 @@ class Node(AbstractNode, LiveParamsMixin):
 
         def log_bundle(bundle: list[str], level: int):
             # When receiving multiple lines of text at ones it is not guaranteed that they belong
-            # together - we don't know if they were generated by a single log call or multiple. 
-            # By looking for our special pattern we can identify new log calls and emit them as 
+            # together - we don't know if they were generated by a single log call or multiple.
+            # By looking for our special pattern we can identify new log calls and emit them as
             # separate messages.
             seg = []
             for line in bundle:
@@ -384,7 +377,7 @@ class Node(AbstractNode, LiveParamsMixin):
 
         try:
             # Since introducing setpgrp above _process.wait hangs indefinitely
-            #self._process.wait(timeout)
+            # self._process.wait(timeout)
             os.waitpid(self.pid, 0)
         except subprocess.TimeoutExpired:
             raise TimeoutError("Node did not shutdown within the specified timeout")
@@ -414,7 +407,7 @@ class Node(AbstractNode, LiveParamsMixin):
         self.logger.info(f"Sending signal {signame} to {repr(self)}")
 
         try:
-            #os.killpg(self.pid, signum)
+            # os.killpg(self.pid, signum)
             self._process.send_signal(signum)
         except ProcessLookupError:
             self.logger.info(
