+++ better_launch/launcher.py
@@ -24,7 +24,7 @@ from rclpy.node import (
     Subscription as RosSubscriber,
 )
 from rclpy.qos import QoSProfile, qos_profile_services_default
-from ament_index_python.packages import get_package_share_directory
+from ament_index_python.packages import get_package_prefix, get_package_share_directory
 
 if TYPE_CHECKING:
     # Surprisingly large imports, so we only import them if we actually need them
@@ -647,6 +647,9 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
         package: str = None,
         filename: str = None,
         subdir: str = "**",
+        *,
+        install_dirs: str | list[str] = "share",
+        append_package_name: bool = True,
     ) -> str:
         """Resolve a path to a file or package.
 
@@ -683,6 +686,9 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
             self.logger.info(f"find({package}, {filename}, {subdir}):1 -> {filename}")
             return filename
 
+        if isinstance(install_dirs, str):
+            install_dirs = [install_dirs]
+
         if not package:
             package, _ = get_package_for_path(os.path.dirname(self.launchfile))
 
@@ -694,11 +700,15 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
                     f"package must be a single name or absolute path ({package})"
                 )
             else:
-                base_path = get_package_share_directory(package)
+                base_path = get_package_prefix(package)
         else:
             raise ValueError(
                 f"find({package}, {filename}, {subdir}): could not determine package"
             )
+
+        base_path = Path(base_path).resolve().absolute()
+        if filename:
+            filename = f"{package}/**/{filename}"
 
         if not filename and subdir in (None, "", "**"):
             self.logger.info(f"find({package}, {filename}, {subdir}):2 -> {base_path}")
@@ -707,28 +717,55 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
         if not subdir:
             subdir = "**"
 
-        for candidate in Path(base_path).glob(subdir):
-            if not filename:
-                # Return the first candidate
-                ret = str(candidate.resolve().absolute())
-                self.logger.info(f"find({package}, {filename}, {subdir}):3 -> {ret}")
-                return ret
-
-            if candidate.is_file() and candidate.match(f"**/{filename}"):
-                # We found a match
-                ret = str(candidate.resolve().absolute())
-                self.logger.info(f"find({package}, {filename}, {subdir}):4 -> {ret}")
-                return ret
-
-            elif candidate.is_dir():
-                # Candidate is a dir, search the filename within
-                ret = next(candidate.glob(f"**/{filename}"), None)
-                if ret:
-                    ret = str(ret.resolve().absolute())
+        def search_base_path(path: Path) -> str:
+            for candidate in path.glob(subdir):
+                if not filename:
+                    # Return the first candidate
+                    ret = str(candidate.resolve().absolute())
                     self.logger.info(
-                        f"find({package}, {filename}, {subdir}):5 -> {ret}"
+                        f"find({package}, {filename}, {subdir}):3 -> {ret}"
                     )
                     return ret
+
+                if candidate.is_file() and candidate.match(f"**/{filename}"):
+                    # We found a match
+                    ret = str(candidate.resolve().absolute())
+                    self.logger.info(
+                        f"find({package}, {filename}, {subdir}):4 -> {ret}"
+                    )
+                    return ret
+
+                elif candidate.is_dir():
+                    # Candidate is a dir, search the filename within
+                    ret = next(candidate.glob(f"**/{filename}"), None)
+                    if ret:
+                        ret = str(ret.resolve().absolute())
+                        self.logger.info(
+                            f"find({package}, {filename}, {subdir}):5 -> {ret}"
+                        )
[…]
