### 2026-02-25 9e8a90a
Added a log message for automatic package resolution in find
