+++ better_launch/launcher.py
@@ -652,10 +652,10 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
 
         If the `filename` is absolute, all other arguments will be ignored and the filename will be returned.
 
-        When `package` names a package discoverable by ament, the corresponding ROS2 package path will be used as the base path. Instead of a package name you may also provide an absolute path, in which case it will become the base path. 
+        When `package` names a package discoverable by ament, the corresponding ROS2 package path will be used as the base path. Instead of a package name you may also provide an absolute path, in which case it will become the base path.
 
         Otherwise, if `package` was not specified we attempt to locate the current launch file's package by searching its directory and parent directories for a `package.xml`. If the package cannot be determined an exception is raised.
-        
+
         `subdir` accepts [glob](https://docs.python.org/3/library/glob.html) patterns and can be used to resolve ambiguities, e.g. `lib/**` (anywhere inside the package's lib folder) or `share/` (directly inside the share folder). If not specified, "**" will be used (any file or directory inside the base path).
 
         If only `subdir` is provided but not `filename`, the first matching candidate is returned. Otherwise the discovered candidates will be searched for the given filename.
@@ -687,6 +687,9 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
 
         if not package:
             package, _ = get_package_for_path(os.path.dirname(self.launchfile))
+            self.logger.warning(
+                f"find: package not provided, resolved {self.launchfile} to {package}"
+            )
 
         if package:
             if os.path.isabs(package):
@@ -703,8 +706,8 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
             )
 
         # In some workspaces, package files are not collected in their own package folders.
-        # Instead, workspace/install has global include, bin, lib, share, etc. folders where 
-        # all the package files are placed, which is quite annoying for us. We fix this by 
+        # Instead, workspace/install has global include, bin, lib, share, etc. folders where
+        # all the package files are placed, which is quite annoying for us. We fix this by
         # requiring the filename to appear after the package name without trying to guess
         # how the package files are organized.
         base_path = Path(base_path).resolve().absolute()
@@ -722,17 +725,13 @@ Please fasten your seatbelts and secure all baggage underneath your chair.
             if not filename:
                 # Return the first candidate
                 ret = str(candidate.resolve().absolute())
-                self.logger.info(
-                    f"find({package}, {filename}, {subdir}):3 -> {ret}"
-                )
+                self.logger.info(f"find({package}, {filename}, {subdir}):3 -> {ret}")
                 return ret
 
             if candidate.is_file() and candidate.match(f"**/{filename}"):
                 # We found a match
                 ret = str(candidate.resolve().absolute())
-                self.logger.info(
-                    f"find({package}, {filename}, {subdir}):4 -> {ret}"
-                )
+                self.logger.info(f"find({package}, {filename}, {subdir}):4 -> {ret}")
                 return ret
 
             elif candidate.is_dir():
