### 2026-02-25 #8958 @danielskatz
Congratulations to @ndahn (Nikolas Dahn) on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you
