### 2026-02-26 89e7361
get_package_for_path should now be more reliable
